Migrating the Android quick start to Android Studio
