+++ client/android/ZUMOAPPNAME/app/src/main/res/values-v11/styles.xml
