+++ client/android/ZUMOAPPNAME/app/src/main/res/values-v14/styles.xml
