+++ client/android/ZUMOAPPNAME/app/src/main/res/values/strings.xml
