+++ client/android/ZUMOAPPNAME/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Jan 11 21:44:32 PST 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.2.1-all.zip
