--- client/android/ZUMOAPPNAME/lint.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<lint>
-</lint>
